requirements.txt
aioredis==1.2.0
amqp==2.5.0
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
billiard==3.6.0.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
dateparser==0.7.1
defusedxml==0.6.0
Django==2.2.4
django-channels==0.7.0
djangorestframework==3.10.2
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
hiredis==1.0.0
httplib2==0.13.1
humanize==0.5.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jira==2.0.0
kombu==4.6.3
librabbitmq==2.0.0
maya==0.6.1
msgpack==0.6.1
oauthlib==3.0.2
pbr==5.4.2
pendulum==2.0.5
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pygsheets==2.0.2
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
pytzdata==2019.2
redis==3.3.6
regex==2019.6.8
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
snaptime==0.2.4
sqlparse==0.3.0
threaded==4.0.3
Twisted==19.7.0
txaio==18.8.1
typing==3.7.4
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zope.interface==4.6.0